refactor: move `load_yaml_dags()` to dagfactory module
