+++ dagfactory/__init__.py
@@ -1,3 +1,2 @@
 """Modules and methods to export for easier access"""
-from .dagfactory import DagFactory
-from .utils import load_yaml_dags
+from .dagfactory import DagFactory, load_yaml_dags
