+++ dagfactory/dagfactory.py
@@ -1,11 +1,16 @@
 """Module contains code for loading a DagFactory config and generating DAGs"""
+import logging
 import os
-from typing import Any, Dict, Optional, Union, List
+from itertools import chain
+from pathlib import Path
+from typing import Any, Dict, List, Optional, Union
 
 import yaml
+from airflow.configuration import conf as airflow_conf
 from airflow.models import DAG
 
 from dagfactory.dagbuilder import DagBuilder
+
 
 # these are params that cannot be a dag name
 SYSTEM_PARAMS: List[str] = ["default", "task_groups"]
@@ -157,3 +162,36 @@ class DagFactory:
             del globals[dag_to_remove]
 
     # pylint: enable=redefined-builtin
+
+
+def load_yaml_dags(
+    globals_dict: Dict[str, Any],
+    dags_folder: str = airflow_conf.get("core", "dags_folder"),
+    suffix=None,
+):
+    """
+    Loads all the yaml/yml files in the dags folder
+
+    The dags folder is defaulted to the airflow dags folder if unspecified.
+    And the prefix is set to yaml/yml by default. However, it can be
+    interesting to load only a subset by setting a different suffix.
+
+    :param globals_dict: The globals() from the file used to generate DAGs
+    :dags_folder: Path to the folder you want to get recursively scanned
+    :suffix: file suffix to filter `in` what files to scan for dags
+    xº"""
+    # chain all file suffixes in a single iterator
+    logging.info("Loading DAGs from %s", dags_folder)
+    if suffix is None:
+        suffix = [".yaml", ".yml"]
+    candidate_dag_files = []
+    for suf in suffix:
+        candidate_dag_files = chain(
+            candidate_dag_files, Path(dags_folder).rglob(f"*{suf}")
+        )
+
+    for config_file_path in candidate_dag_files:
+        config_file_abs_path = str(config_file_path.absolute())
+        DagFactory(config_file_abs_path).generate_dags(globals_dict)
+        logging.info("DAG loaded: %s", config_file_path)
+
